added new and updated servers, timetables and html
the python server code should be able to handle multiple servers now, an upgrade from just 2. i have tested with 6 servers, they all communicated properly. i'm pretty sure it could handle up to 20(the maximum needed for the project).
see the script again to start the server,
tt- files are the timetables for the respective stations.
html is the client website.
s1.py is the python server code.